Create Habit Model interfaces
Our project will follow a Model-View-ViewModel architecture as it
completely decouples the View from the Model, making our code more
testable (we'll have to do that later, and it would be painful if we
follow MVC).

This commit adds an interface for `Habit` which we'll implement as an
Entity in an SQLite database for now, but will also be implemented for
Firebase.

It also adds a `HabitRepository` class for interacting with the data
store (SQLite or Firebase). For now, it only has functions
`getAllHabits` and `insert` because this is what's relevant for
adding habits (#1).