Pass Habit to ViewHabitFragment
I first thought about using tags to pass data but since we're already
using ViewModels, it would be more consistent to use them.

We first need to find a way to select a `Habit` to view. This is done by
creating an 'onClick' interface for the `HabitAdapter` which is
implemented by the `TodaysHabitsFragment`. The `TodaysHabitsFragment`
selects the habit while the `ViewHabitFragment` observes for changes to
the `LiveData` and updates the UI accordingly.