Reuse `NewHabitFragment` for editing a Habit
Change `NewHabitFragment` to `CreateEditHabitFragment`
It just made more sense to use the same views for creating the
component as editing it, so that the View Habit fragment has less
responsibilities (displaying its information, displaying habit events in
the future AND editing?).

This is done with some LiveData gymnastics. Maybe I should just
implement everything here since it's not really in the scope of this
course and more of an extra mile thing.