Fix Back button navigation issues
Fixed navigation issues by overriding the onNavigateUp event.
AllHabits now always return to TodaysHabits.

ViewHabit can either return to TodaysHabits or AllHabits, depending on
where the habit was clicked. To keep track of this, we created a new
ViewModel called PreviousListViewModel that both List Fragments update
when a habit is clicked for viewing. This int value is then used by the
ViewHabit fragment for navigation.