Testing: Provide example of ViewModel testing
Provided an example on how to test the ViewModel classes.
Dependency injection makes unit testing straightforward.

In this example, HabitViewModel only depends on one other class, which
is the HabitRepository. We only need to verify that:

1. HabitViewModel's methods call the right methods from HabitRepository
2. It passes the right inputs to the method

We first create variables for the parameters since we'll need references
to them to verify the called method.
We then call the method that we need to test with the created
parameters. In this case, it's `HabitViewModel.insert` method.
Then, we verify that the mock `HabitRepository.insert` method that the
`HabitViewModel.insert` is supposed to call was indeed called, along
with the right parameters. This is done by capturing the method
parameters with `ArgumentCaptor`s.

Testing the MutableLiveData `mSelectedHabit` might be a bit more
complicated, but the rest should follow the same pattern as the
`testInsert` case.

Summary:

1. Call method to test with example parameters
2. Verify the number of invocations using `verify(MOCK,
times(N_INVOCATIONS)).METHOD(CAPTORS)`
3. Capture the values passed to the mock
4. Compare the captor values to the passed parameters